Part DAO/MVC Fronted integration - Part 3 (Educateur is OK)

+++ views/educateurs/homeView.php
@@ -2,15 +2,18 @@
     <div class="card-body p-4">
         <div class="d-flex justify-content-between">
             <h5 class="card-title fw-semibold mb-4">Educateurs</h5>
-            <a class="btn btn-primary m-1" href="index.php?page=licencies&action=add">Ajouter</a>
+            <a class="btn btn-primary m-1" href="index.php?page=educateurs&action=add">Ajouter</a>
         </div>
+        <!-- Flash message section -->
+        <?php include './views/components/flashMsg.php'; ?>
+        <!-- Records -->
         <div class="table-responsive">
             <?php if (count($educateurs) > 1) : ?>
                 <table class="table text-nowrap mb-0 align-middle">
                     <thead class="text-dark fs-4">
                         <tr>
                             <th class="border-bottom-0">
-                                <h6 class="fw-semibold mb-0">No licence</h6>
+                                <h6 class="fw-semibold mb-0">N° licence</h6>
                             </th>
                             <th class="border-bottom-0">
                                 <h6 class="fw-semibold mb-0">Nom</h6>
@@ -34,37 +37,39 @@
                     </thead>
                     <tbody>
                         <?php foreach ($educateurs as $educateur) : ?>
-                            <tr>
-                                <td class="border-bottom-0">
-                                    <h6 class="fw-semibold mb-0"><?= $educateur->getNumLicence() ?></h6>
-                                </td>
-                                <td class="border-bottom-0">
-                                    <h6 class="fw-semibold mb-0"><?= $educateur->getNom() ?></h6>
-                                </td>
-                                <td class="border-bottom-0">
-                                    <h6 class="fw-semibold mb-1"><?= $educateur->getPrenom() ?></h6>
-                                </td>
-                                <td class="border-bottom-0">
-                                    <h6 class="fw-semibold mb-1"><?= $educateur->getIdCategorie() ?></h6>
-                                </td>
-                                <td class="border-bottom-0">
-                                    <h6 class="fw-semibold mb-1"><?= $educateur->getEmail() ?></h6>
-                                </td>
-                                <td class="border-bottom-0">
-                                    <h6 class="fw-semibold mb-1">
-                                        <?php if ($educateur->isAdmin()) : ?>
-                                            <span class="badge bg-success rounded-3 fw-semibold">Oui</span>
-                                        <?php else : ?>
-                                            <span class="badge bg-danger rounded-3 fw-semibold">Non</span>
-                                        <?php endif; ?>
-                                    </h6>
-                                </td>
-                                <td class="border-bottom-0 d-flex">
-                                    <a class="btn btn-primary m-1" href="index.php?page=educateurs&action=show&id=<?= $educateur->getId(); ?>">Voir</a>
-                                    <a class="btn btn-warning m-1" href="index.php?page=educateurs&action=edit&id=<?= $educateur->getId(); ?>">Modifier</a>
-                                    <a class="btn btn-danger m-1" href="index.php?page=educateurs&action=delete&id=<?= $educateur->getId(); ?>">Supprimer</a>
-                                </td>
-                            </tr>
+                            <?php if ($educateur->getId() != 1) : ?>
+                                <tr>
+                                    <td class="border-bottom-0">
+                                        <h6 class="fw-semibold mb-0"><?= $educateur->getNumLicence() ?></h6>
+                                    </td>
+                                    <td class="border-bottom-0">
+                                        <h6 class="fw-semibold mb-0"><?= $educateur->getNom() ?></h6>
+                                    </td>
+                                    <td class="border-bottom-0">
+                                        <h6 class="fw-semibold mb-1"><?= $educateur->getPrenom() ?></h6>
+                                    </td>
+                                    <td class="border-bottom-0">
+                                        <h6 class="fw-semibold mb-1"><?= $educateur->getIdCategorie() ?></h6>
+                                    </td>
+                                    <td class="border-bottom-0">
+                                        <h6 class="fw-semibold mb-1"><?= $educateur->getEmail() ?></h6>
+                                    </td>
+                                    <td class="border-bottom-0">
+                                        <h6 class="fw-semibold mb-1">
+                                            <?php if ($educateur->isAdmin()) : ?>
+                                                <span class="badge bg-success rounded-3 fw-semibold">Oui</span>
+                                            <?php else : ?>
+                                                <span class="badge bg-danger rounded-3 fw-semibold">Non</span>
+                                            <?php endif; ?>
+                                        </h6>
+                                    </td>
+                                    <td class="border-bottom-0 d-flex">
+                                        <a class="btn btn-primary m-1" href="index.php?page=educateurs&action=show&id=<?= $educateur->getId(); ?>">Voir</a>
+                                        <a class="btn btn-warning m-1" href="index.php?page=educateurs&action=edit&id=<?= $educateur->getId(); ?>">Modifier</a>
+                                        <a class="btn btn-danger m-1" href="index.php?page=educateurs&action=delete&id=<?= $educateur->getId(); ?>">Supprimer</a>
+                                    </td>
+                                </tr>
+                            <?php endif; ?>
                         <?php endforeach; ?>
                     </tbody>
                 </table>
